+++ cogno_gateway/__init__.py
@@ -39,6 +39,11 @@ from cogno_gateway.types import (
 from cogno_gateway.ports import Channel, GatewayError, PresenceAwareChannel
 from cogno_gateway.chunker import split_message
 from cogno_gateway.markup import to_channel_markup
+from cogno_gateway.net import (
+    FamilyFallbackBackend,
+    FamilyFallbackTransport,
+    build_async_client,
+)
 from cogno_gateway.web import WebChannel
 from cogno_gateway.telegram import TelegramChannel
 from cogno_gateway.evolution import EvolutionChannel
@@ -77,6 +82,9 @@ __all__ = [
     "GatewayError",
     "split_message",
     "to_channel_markup",
+    "build_async_client",
+    "FamilyFallbackBackend",
+    "FamilyFallbackTransport",
     "WebChannel",
     "TelegramChannel",
     "EvolutionChannel",